Merged PR 1400: Initial Solution migration step
Initial Solution migration step

This will load a solution if that is all that is available. The step will ask a few questions and gather results to identify the order.

NOTE: No references will be found with an old style csproj file, so it will not actually find an order. However, once it's converted it will be able to. Subsequent PRs will address this bootstrapping issue.

Related work items: #1759, #1815