Merged PR 1402: Set VSToolsPath for MSBuildWorkspace properties
Set VSToolsPath for MSBuildWorkspace properties

This does a search similar to the MSBuildLocator but searches for VS instances. Unfortunately, the .NET 5 app selects the netcoreapp2.0 target from that library which skips over any of the VS installs. Since we just need it for targets files (at least at this point), we set it so it will resolve properly.

Related work items: #1821