Merged PR 1426: Show base directory at app startup
Since users may be adding their own extensions next to the migrator app, I want to make sure they know where to find it. This may not be obvious, especially if they install it as a CLI tool.

After trying a couple approaches (like custom commands to show config, etc.), I think the best solution is just this one-line change to show the app context base directory that the app probes from as part of startup.