Merged PR 1430: Track current project via IMigrationContext
This change tracks completion of migration of projects so that it can be persisted to disk. This is done by a file named `.aspnetmigrator` in the same directory as the input file.  At the moment, this just tracks completed projects, but more required state could be added.

Related work items: #1807, #1837, #1838, #1839