Add initial reporting from portability service (#22)
This adds some infrastructure to gather the referenced APIs and submit that to the portability service. This does not attempt to categorize them, but rather to see what we can do with what's currently available. There is an initial implementation of checking if an API is available, which we'll want to update to react to things such as if a user is going to move to Linux or stay on Windows.