Change TFM comparer to see framework as before core/standard
After identifying the most likely TFM, we check if we will be downgrading the current project. If a project was selected to target .NET Standard but was currently targeting .NET 4.7.2 (or any .NET framework that supports .NET Standard), it would be marked as downgrading and select .NET 4.7.2 as the final TFM. This special cases .NET Standard so that it will always be seen as after .NET Framework TFMs.