Replace ITargetTFMSelector with a filter-based pattern
This change doesn't change too much for callers, but changes how a TFM is selected. Previously, one class held all the knowledge about how to select a TFM for a project. This swaps that out for a filter chain that can attempt to change the final result. Each of the filters can do whatever checks they may want, but the ITargetFrameworkSelectorFilter.Builder that is passed through will only update the TFM if it is a higher or more constrained TFM.