Remove central version management (#573)
This seems to be causing problems with the way we're building our
extensions. Somehow it's detecting that we have a version when no
PackageReferences do. Since this is a preview feature, it appears we may
have hit a bug with that. For now, swapping back to non-centrally
managed versions allows for a more seamless inner dev-loop.