Separate out web related items into its own extension (#582)
This removes most of the web-specific info to a separate extension so that it can have web-related things co-located with it. The analyzers/codefixers are not self-contained, but that may require a bit more thinking as we distribute them. This is a first step towards organizing extensions by technology areas.

The resulting output looks like this:

```
upgrade-assistant
|- Microsoft.DotNet.UpgradeAssistant.Cli.dll
|- ....
|- extensions
  |- default
    |- [files]
  |- web
    |- [files]
```