Redirect bin/obj folders for extensions to minimize file write races
When building on the CI, there are some race conditions that occur
periodically to write files to disk. This is because the extensions are
built separately for their own projects and for when they are being
included in the CLI. This change sets a custom OutDir and
IntermediateOutputPath when needed so there are separate places these
get written.