Move UpgradeOptions into CLI class
UpgradeOptions should be an implementation detail of the CLI as it is tightly bound to the command line parsing. The names on it cannot be changed without breaking that binding, and has ended up being a bag of a bunch of options.

This change introduces new options for the parts of the system that needed it and flows the values to them via:

- Directly configuring the options in the ConfigureServices
- Setting them up as extension options for values that are needed in extensions and cannot be directly set