Add implementation of extension creation/updating/adding/restoring (#756)
This change adds a NuGet backed way of creating/adding/restoring extensions.

With this change, you can add/remove/list/use extensions. However, in order to restore on demand things that are currently synchronous need to be made async. That will be done in a subsequent change to minimize the amount of churn.