Use .netstandard2.0 for default extension (#1026)
We want to use .NET Standard for the extensions if possible so that we
could potentially run them in Visual Studio at some point. The reason
the default extension was not .NET Standard is because the try-convert
stuff uses MSBuild directly and was forcing .NET 5+ to be the TFM. This
change moves try-convert to be its own extension which allows the
default extension to be .NET Standard.