Update System.Configuration.ConfigurationManager
dotnet/msbuild#7629 updated MSBuild's reference to
`System.Configuration.ConfigurationManager` to the current 6.0.0.0
version. That version is available in the SDK directory and
MSBuildLocator tries to load it but fails. Figuring out why is tracked
as microsoft/MSBuildLocator#159, but this works around the problem by
updating to (and thus redistributing) the same version that MSBuild now
references, which is compatible with older SDKs/MSBuilds as well.