Prerelease NuGet for Integration.Tests
Preview versions of Visual Studio come with preview versions
of the .NET SDK, which have new MSBuild and NuGet versions.
This can cause problems because the version will be newer
than the one directly referenced by this test. Change the
test project's reference to allow prerelease NuGet packages,
so that it pulls the very latest NuGet, which should almost
always be a higher version than the one in the SDK used in
the tests.