Fixed memory corruption in VisualStudioFinder
VisualStudioFinder.GetLatestPath() uses COM to iterate over the
install locations for Visual Studio but somehow, once the method
returns to the caller (GetLatestVisualStudioPath), none of the
methods can be relied on to return sane values.

Therefore, the easiest solution is to track the values we care about
and just return them to our caller who can then avoid needing to re-make
the same COM invocations a second time.