Timestamp NuGet packages
Several people have been bitten while testing the UpgradeAssistant with
local builds of the Mappings nuget package because every build produces
the same versioned package and if they forget to delete the cached version
in the NuGet cache directory, then their changes will appear to not take
affect because UA is using an old version of their changes.

Also modified the Nerdbank.GitVersioning version.json file to use a base of
"1.0" instead of "1.0.0".